refactor: move stream metadata dedupe helper to sibling module
